Refactor the LanguageService into a separate module
- Removed builder parts from typescript.ts and moved them to build.ts
- Made renamed TypeScriptBuilder to LanguageServiceContext
- Changed Language Service functions to accept LS context as parameter
- Made use of project-cache easier by providing sane defaults
